requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.0.12
dj-config-url==0.1.1
Django==3.2.13
django-restframework==0.0.1
djangorestframework==3.13.1
googlemaps==4.6.0
gunicorn==20.1.0
idna==3.3
numpy==1.21.6
pandas==1.3.5
Pillow==9.1.1
psycopg2-binary==2.9.2
PyJWT==2.4.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
stripe==3.4.0
twilio==7.9.3
typing_extensions==4.2.0
urllib3==1.26.9
whitenoise==6.2.0
